// typings/dataWrangler.d.ts
export interface IDataWranglerExtensionAPI {
    /**
     * Promise indicating whether all parts of the extension have completed loading or not.
     * @type {Promise<void>}
     * @memberof IDataWranglerExtensionAPI
     */
    ready: Promise<void>;

    /**
     * Launches the Data Wrangler using the given variable name.
     */
    launchDataWranglerUsingVariable: (
        notebookUri: Uri,
        variableName: string,
        session: {
            metadata: KernelConnectionMetadata;
            connection: IKernelConnectionInfo;
        },
        exportLocation?: {
            outputId?: string;
        }
    ) => Promise<void>;
}
// ...
export type LocalKernelSpecConnectionMetadata = Readonly<{
    kernelModel?: undefined;
    kernelSpec: IJupyterKernelSpec;
    /**
     * Indicates the interpreter that may be used to start the kernel.
     * If possible to start a kernel without this Python interpreter, then this Python interpreter will be used for intellisense & the like.
     * This interpreter could also be the interpreter associated with the kernel spec that we are supposed to start.
     */
    interpreter?: PythonEnvironment;
    kind: 'startUsingLocalKernelSpec';
    id: string;
}>;
/**
 * Connection metadata for Remote Kernels started using kernelspec (JSON).
 * This could be a raw kernel (spec might have path to executable for .NET or the like).
 * If the executable is not defined in kernelspec json, & it is a Python kernel, then we'll use the provided python interpreter.
 */
export type RemoteKernelSpecConnectionMetadata = Readonly<{
    kernelModel?: undefined;
    interpreter?: undefined;
    kernelSpec: IJupyterKernelSpec;
    kind: 'startUsingRemoteKernelSpec';
    baseUrl: string;
    id: string;
}>;
// ...
export type PythonKernelConnectionMetadata = Readonly<{
    kernelSpec: IJupyterKernelSpec;
    interpreter: PythonEnvironment;
    kind: 'startUsingPythonInterpreter';
    id: string;
}>;
// ...
export type LiveRemoteKernelConnectionMetadata = Readonly<{
    kernelModel: LiveKernelModel;
    /**
     * Python interpreter will be used for intellisense & the like.
     */
    interpreter?: PythonEnvironment;
    baseUrl: string;
    kind: 'connectToLiveRemoteKernel';
    id: string;
}>;

export type KernelConnectionMetadata =
    | LocalKernelSpecConnectionMetadata
    | RemoteKernelSpecConnectionMetadata
    | PythonKernelConnectionMetadata
    | LiveRemoteKernelConnectionMetadata;
export type ActiveKernel = LiveRemoteKernelConnectionMetadata;

export interface IKernelSocket {
    /**
     * Whether the kernel socket is read & available for use.
     * Use `onDidChange` to be notified when this changes.
     */
    ready: boolean;
    /**
     * Event fired when the underlying socket state changes.
     * E.g. when the socket is connected/available or changes to another socket.
     */
    onDidChange: Event<void>;
    /**
     * Sends data to the underlying Jupyter kernel over the socket connection.
     * This bypasses all of the jupyter kernel comms infrastructure.
     */
    sendToRealKernel(data: any, cb?: (err?: Error) => void): void;
    /**
     * Adds a listener to a socket that will be called before the socket's onMessage is called. This
     * allows waiting for a callback before processing messages
     */
    addReceiveHook(hook: (data: WebSocketData) => Promise<void>): void;
    /**
     * Removes a listener for the socket. When no listeners are present, the socket no longer blocks
     */
    removeReceiveHook(hook: (data: WebSocketData) => Promise<void>): void;
    /**
     * Adds a hook to the sending of data from a websocket. Hooks can block sending so be careful.
     */
    addSendHook(hook: (data: any, cb?: (err?: Error) => void) => Promise<void>): void;
    /**
     * Removes a send hook from the socket.
     */
    removeSendHook(hook: (data: any, cb?: (err?: Error) => void) => Promise<void>): void;
}

export type IKernelConnectionInfo = {
    /**
     * Gives access to the jupyterlab Kernel.IKernelConnection object.
     */
    connection: Kernel.IKernelConnection;
    /**
     * Underlying socket used by jupyterlab/services to communicate with kernel.
     * See jupyterlab/services/kernel/default.ts
     */
    kernelSocket: IKernelSocket;
};